LPD-20917 Refactor, hoist isSPAEnabled to the top level utils for usage anywhere
