+++ modules/test/playwright/tests/frontend-js-spa-web/main/settings.spec.ts
@@ -8,8 +8,8 @@ import {expect, mergeTests} from '@playwright/test';
 import {isolatedLayoutTest} from '../../../fixtures/isolatedLayoutTest';
 import {loginTest} from '../../../fixtures/loginTest';
 import {systemSettingsPageTest} from '../../../fixtures/systemSettingsPageTest';
+import isSPAEnabled from '../../../utils/isSPAEnabled';
 import {waitForAlert} from '../../../utils/waitForAlert';
-import isSPAEnabled from './utils/isSPAEnabled';
 
 export const test = mergeTests(
 	isolatedLayoutTest({publish: false}),
